Move GameService code to GameManagementService
