+++ src/main/java/com/rocketden/main/controller/v1/GameController.java
@@ -3,7 +3,7 @@ package com.rocketden.main.controller.v1;
 import com.rocketden.main.dto.room.RoomDto;
 import com.rocketden.main.dto.game.StartGameRequest;
 
-import com.rocketden.main.service.GameService;
+import com.rocketden.main.service.GameManagementService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,10 +15,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class GameController extends BaseRestController {
 
-    private final GameService service;
+    private final GameManagementService service;
 
     @Autowired
-    public GameController(GameService service) {
+    public GameController(GameManagementService service) {
         this.service = service;
     }
