+++ src/main/java/com/rocketden/main/service/GameManagementService.java
@@ -7,7 +7,12 @@ import java.util.Map;
 
 import com.rocketden.main.dao.RoomRepository;
 import com.rocketden.main.dto.game.GameDto;
+import com.rocketden.main.dto.game.StartGameRequest;
 import com.rocketden.main.dto.notification.NotificationDto;
+import com.rocketden.main.dto.room.RoomDto;
+import com.rocketden.main.dto.room.RoomMapper;
+import com.rocketden.main.exception.RoomError;
+import com.rocketden.main.exception.api.ApiException;
 import com.rocketden.main.game_object.Game;
 import com.rocketden.main.game_object.GameNotification;
 import com.rocketden.main.game_object.Player;
@@ -26,10 +31,11 @@ public class GameManagementService {
     private final LiveGameService liveGameService;
     private final NotificationService notificationService;
     private final SubmitService submitService;
-    private Map<String, Game> currentGameMap;
+    private final Map<String, Game> currentGameMap;
 
     @Autowired
-    protected GameManagementService(RoomRepository repository, SocketService socketService, LiveGameService liveGameService, NotificationService notificationService, SubmitService submitService) {
+    protected GameManagementService(RoomRepository repository, SocketService socketService, LiveGameService liveGameService,
+                                    NotificationService notificationService, SubmitService submitService) {
         this.repository = repository;
         this.socketService = socketService;
         this.liveGameService = liveGameService;
@@ -38,19 +44,45 @@ public class GameManagementService {
         currentGameMap = new HashMap<>();
     }
 
-    public Game createAddGameFromRoom(Room room) {
-        // TODO: Create the game from the room (or roomId).
-        Game game = new Game();
-        currentGameMap.put(room.getRoomId(), game);
-        return null;
-    }
-
     protected Game getGameFromRoomId(String roomId) {
         return currentGameMap.get(roomId);
     }
 
     protected void removeGame(String roomId) {
         currentGameMap.remove(roomId);
+    }
+
+    // When host starts the game, redirect everyone and initialize the game state
+    public RoomDto startGame(String roomId, StartGameRequest request) {
+        Room room = repository.findRoomByRoomId(roomId);
+
+        // If requested room does not exist in database, throw an exception.
+        if (room == null) {
+            throw new ApiException(RoomError.NOT_FOUND);
+        }
+
+        // If user making request is not the host, throw an exception.
+        if (!request.getInitiator().getNickname().equals(room.getHost().getNickname())) {
+            throw new ApiException(RoomError.INVALID_PERMISSIONS);
+        }
+
+        room.setActive(true);
+        repository.save(room);
+
+        // Initialize game state
+        createAddGameFromRoom(room);
+
+        RoomDto roomDto = RoomMapper.toDto(room);
+        socketService.sendSocketUpdate(roomDto);
+        return roomDto;
+    }
+
+    // Initialize and add a game object from a room object
+    public Game createAddGameFromRoom(Room room) {
+        // TODO: Create the game from the room (or roomId).
+        Game game = new Game();
+        currentGameMap.put(room.getRoomId(), game);
+        return null;
     }
 
     // Test the submission and return a socket update.
